+++ projects/cli/src/Lunapack.Cli/Properties/AssemblyInfo.cs
@@ -1,4 +1,5 @@
 [assembly: System.Runtime.CompilerServices.InternalsVisibleTo("Lunapack.Cli.UnitTests")]
+[assembly: System.Runtime.CompilerServices.InternalsVisibleTo("Lunapack.Cli.SecurityTests")]
 [assembly: System.Diagnostics.CodeAnalysis.SuppressMessage(
     "Usage",
     "MA0004:Use Task.ConfigureAwait(false)",